♻️ refactor: migrate Library multi-select to BookMultiSelectViewModel + extract BookSelectionScaffold
